+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # Diskover Change Log
 
-## [1.5.0-rc9] = 2018-06-04
+## [1.5.0-rc9] = 2018-06-05
 ### notice
 - diskover-bot-launcher.sh has been updated, when updating with git please check that any of your env settings at top of file have not changed, you may need to edit these again
 - if using the autotag flag, you may want to add a new custom tag in diskover-web admin page for "autotag" if you are using that as the tag_custom value in autotag patterns
@@ -13,9 +13,11 @@
 - \ (backslash) char to escape_chars function
 - v1.3 of diskover-bot-launcher.sh - added restart bot cli arg -r (changed redis worker remove to -R) and cleaned up script
 - improved dupe checking
+- better killredisconn.py (output of status and -f arg to force remove)
 ### fixed
 - bug with dupe md5 check
 - bug with regular expression matching for directory excludes
+- bug with killredisconn.py not working with Python 3
 
 ## [1.5.0-rc8] = 2018-05-26
 ### notice